fixed error on measuring due to missing floating point arithmetic

+++ ArduinoLoRaSensorBox/src/LightSensor.h
@@ -20,7 +20,7 @@ public:
     uint16_t getResistance();
 private:
     uint8_t _pin;
-    uint16_t _raw;
+    float _raw;
     float _rsensor;
     uint32_t _n;
 };
